+++ platformio/builder/scripts/espressif.py
@@ -92,7 +92,7 @@ env.Replace(
         "-cp", "$UPLOAD_PORT",
         "-cf", "$SOURCE"
     ],
-    UPLOADCMD='$UPLOADER $UPLOADERFLAGS',
+    UPLOADCMD='"$UPLOADER" $UPLOADERFLAGS',
 
     PROGNAME="firmware",
     PROGSUFFIX=".elf"
@@ -104,9 +104,10 @@ env.Append(
     BUILDERS=dict(
         ElfToBin=Builder(
             action=" ".join([
-                "$UPLOADER",
-                "-eo", join("$PLATFORMFW_DIR", "bootloaders",
-                            "eboot", "eboot.elf"),
+                '"$UPLOADER"',
+                "-eo",
+                '"%s"' % join("$PLATFORMFW_DIR", "bootloaders",
+                              "eboot", "eboot.elf"),
                 "-bo", "$TARGET",
                 "-bm", "dio",
                 "-bf", "${BOARD_OPTIONS['build']['f_cpu'][:2]}",
@@ -147,7 +148,7 @@ if "FRAMEWORK" in env:
                     "-i", "$UPLOAD_PORT",
                     "-f", "$SOURCE"
                 ],
-                UPLOADCMD='$UPLOADEROTA $UPLOADERFLAGS'
+                UPLOADCMD='"$UPLOADEROTA" $UPLOADERFLAGS'
             )
     except socket.error:
         pass
@@ -163,7 +164,7 @@ else:
         BUILDERS=dict(
             ElfToBin=Builder(
                 action=" ".join([
-                    "$UPLOADER",
+                    '"$UPLOADER"',
                     "-eo", "$SOURCES",
                     "-bo", "${TARGETS[0]}",
                     "-bm", "qio",