+++ HISTORY.rst
@@ -18,6 +18,8 @@ PlatformIO 2.0
   (`issue #379 <https://github.com/platformio/platformio/issues/379>`_)
 * Fixed reset method for Espressif NodeMCU (ESP-12E Module)
   (`issue #380 <https://github.com/platformio/platformio/issues/380>`_)
+* Fixed issue with code builder when build path contains spaces
+  (`issue #387 <https://github.com/platformio/platformio/issues/387>`_)
 
 
 2.6.0 (2015-12-15)